+++ src/features/bumpkins/components/AchievementBadges.tsx
@@ -3,7 +3,7 @@ import { getKeys } from "features/game/types/craftables";
 import { Bumpkin } from "features/game/types/game";
 import { ITEM_DETAILS } from "features/game/types/images";
 
-export const AchievementsPreivew: React.FC<{
+export const AchievementBadges: React.FC<{
   achievemenets?: Bumpkin["achievements"];
 }> = ({ achievemenets = {} }) => {
   const badges = getKeys(achievemenets).map((name) => {
